backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Celery
celery[redis]==5.3.6
redis==5.0.1

# PDF Processing
pymupdf==1.23.21
pikepdf==8.11.2
reportlab==4.0.9

# Image Processing
Pillow==10.2.0

# PPTX
python-pptx==0.6.23

# Google Cloud
google-cloud-vision==3.5.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# AWS S3 (optional)
boto3==1.34.25
